Extract BucketApiImpl to different file
